// backend/src/wsServer/models/messageTypes.ts
export enum MessageType {
	PING = "PING",
	AUTH = "AUTH",
	ID = "ID",
	HISTORY_ADD = "HISTORY_ADD",
	HISTORY_DELETE = "HISTORY_DELETE",
	BOOKMARKS_ADD = "BOOKMARKS_ADD",
	BOOKMARKS_DELETE = "BOOKMARKS_DELETE",
	BOOKMARKS_UPDATE = "BOOKMARKS_UPDATE",
	BOOKMARKS_MOVE = "BOOKMARKS_MOVE",
	TABS_ADD = "TABS_ADD"
}